move status checker myself and child routes
