--- frontend/app/routes/$lang/_public/status/child/_route.tsx
@@ -1,9 +1,0 @@
-import { Outlet } from '@remix-run/react';
-
-/**
- * Do-nothing parent route.
- * (placeholder for future code)
- */
-export default function Route() {
-  return <Outlet />;
-}
