+++ frontend/app/routes/$lang/page-ids.json
@@ -111,12 +111,8 @@
     },
     "status": {
       "index": "CDCP-STAT-0001",
-      "myself": {
-        "index": "CDCP-STAT-0002"
-      },
-      "child": {
-        "index": "CDCP-STAT-0003"
-      }
+      "myself": "CDCP-STAT-0002",
+      "child": "CDCP-STAT-0003"
     },
     "unableToProcessRequest": "CDCP-UNAB-0001",
     "notFound": "CDCP-ERR-0404"
